requirements.txt
mysql-connector-python==8.0.33
Pillow==10.0.0
protobuf==3.20.3
